+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - fraudType parameter to PixInfraction resource
+- bacenId attribute to PixChargeback resource
 
 ## [0.9.0] - 2025-07-08
 ### Added